add sunspot number with silso diff page and api

+++ api/routers/sunspot_number_with_silso.py
@@ -7,6 +7,7 @@ from pydantic import BaseModel
 
 from api.libs import sunspot_number_with_silso, utils
 from api.libs.sunspot_number_with_silso_config import (
+    SunspotNumberDiff,
     SunspotNumberRatio,
     SunspotNumberScatter,
     SunspotNumberWithSilso,
@@ -293,6 +294,67 @@ def save_ratio(body: draw.SaveBody) -> draw.SaveRes:
         json_data = json.load(f_factor_r2)
         factor = json_data["factor"]
     fig = sunspot_number_with_silso.draw_ratio(df, factor, config)
+    fig.savefig(
+        output_path,
+        format=body.format,
+        dpi=body.dpi,
+        bbox_inches="tight",
+        pad_inches=0.1,
+    )
+    return draw.SaveRes(output=str(output_path))
+
+
+@router.get("/draw/diff", response_model=draw.PreviewRes)
+def draw_diff(query: draw.PreviewQuery = Depends()) -> draw.PreviewRes:
+    input_path = Path(query.filename)
+    if not input_path.exists():
+        raise HTTPException(
+            status_code=404, detail=f"file {input_path} not found"
+        )
+    config_path = Path(query.config_name)
+    if not config_path.exists():
+        raise HTTPException(
+            status_code=404, detail=f"config {config_path} not found"
+        )
+    try:
+        with config_path.open("r") as f_config:
+            config = SunspotNumberDiff(**json.load(f_config))
+    except ValueError as e:
+        raise HTTPException(
+            status_code=400, detail=f"config {config_path} is broken"
+        ) from e
+    df = pl.read_parquet(input_path)
+    fig = sunspot_number_with_silso.draw_diff(df, config)
+    img = utils.fig_to_base64(fig)
+    return draw.PreviewRes(img=img)
+
+
+@router.post("/draw/diff", response_model=draw.SaveRes)
+def save_diff(body: draw.SaveBody) -> draw.SaveRes:
+    input_path = Path(body.input)
+    if not input_path.exists():
+        raise HTTPException(
+            status_code=404, detail=f"file {input_path} not found"
+        )
+    config_path = Path(body.config)
+    if not config_path.exists():
+        raise HTTPException(
+            status_code=404, detail=f"config {config_path} not found"
+        )
+    output_path = input_path.with_name(f"diff.{body.format}")
+    if not body.overwrite and output_path.exists():
+        raise HTTPException(
+            status_code=400, detail=f"file {output_path} already exists"
+        )
+    try:
+        with config_path.open("r") as f:
+            config = SunspotNumberDiff(**json.load(f))
+    except ValueError as e:
+        raise HTTPException(
+            status_code=400, detail=f"config {config_path} is broken"
+        ) from e
+    df = pl.read_parquet(input_path)
+    fig = sunspot_number_with_silso.draw_diff(df, config)
     fig.savefig(
         output_path,
         format=body.format,
